requirements.txt
pandas>=1.3.0
numpy>=1.20.0
scipy>=1.7.0
Flask>=2.2
waitress>=2.1
Pillow>=9.0.0
reportlab>=3.6.0
